requirements.txt
python-pptx==0.6.22
google-generativeai==0.3.2
python-dotenv==1.0.1
tqdm==4.66.1
Pillow==9.5.0
requests==2.31.0
jsonschema==4.21.1